+++ src/ciceroscm/component_factory_functions.py
@@ -6,7 +6,7 @@ from .two_layer_ocean import TwoLayerOceanModel
 from .upwelling_diffusion_model import UpwellingDiffusionModel
 
 
-def create_carbon_cycle_model(model_type, pamset):
+def create_carbon_cycle_model(model_type, pamset, pamset_carbon=None):
     """
     Create a CarbonCycleModel instance (factory function).
 
@@ -28,9 +28,9 @@ def create_carbon_cycle_model(model_type, pamset):
         If an undefined carbon cycle model_type is called for
     """
     if model_type == "default":
-        return DefaultCarbonCycleModel(pamset)
+        return DefaultCarbonCycleModel(pamset, pamset_carbon)
     if model_type == "box":
-        return BoxCarbonCycleModel(pamset)
+        return BoxCarbonCycleModel(pamset, pamset_carbon)
     raise ValueError(f"Unknown model type: {model_type}")