+++ src/ciceroscm/parallel/_configdistro.py
@@ -8,6 +8,8 @@ import logging
 import numpy as np
 from scipy.stats import qmc
 
+from ..carbon_cycle.carbon_cycle_mod import CARBON_CYCLE_MODEL_REQUIRED_PAMSET
+
 LOGGER = logging.getLogger(__name__)
 
 # TODO: sensible standard priors for q-terms
@@ -19,63 +21,28 @@ prior_flat = {
     "beto": [0, 7],
     "lambda": [2 / 3.71, 5 / 3.71],
     "mixed": [25, 125],
+    "ocean_efficacy": [0.8, 1.2],
     "qbmb": [0, 2],
     "qo3": [0.4, 0.6],
-    "qdirso2": [-0.55, -0.2],
-    "qindso2": [-1.5, -0.5],
-    "qbc": [0.1, 0.2],
-    "qoc": [-0.1, -0.06],
+    "qdirso2": [-0.006, -0.001],
+    "qindso2": [-0.03, -0.01],
+    "qbc": [0.004, 0.05],
+    "qoc": [-0.008, -0.001],
     "qh2o_ch4": [0.08, 0.1],
-    "aerosol_total": [-3.0, -0.5],
     "beta_f": [0.110, 0.465],
     "mixed_carbon": [25, 125],
+    "ml_w_sigmoid": [2.0, 4.0],
+    "ml_fracmax": [0.3, 0.8],
+    "ml_t_half": [-0.5, 1.0],
+    "npp0": [55, 75],
+    "t_half": [0.3, 0.8],
+    "w_sigmoid": [5, 10],
+    "t_threshold": [3, 7],
+    "w_threshold": [5, 10],
+    "solubility_sens": [0.01, 0.03],
+    "solubility_limit": [0.4, 0.8],
 }
 """dict: Containing a default prior for parameters """
-
-prior_flat_array = np.array(
-    [
-        [5, 25],
-        [0.06, 0.8],
-        [0.161, 0.569],
-        [0.55, 2.55],
-        [0, 7],
-        [2 / 3.71, 5 / 3.71],
-        [25, 125],
-        [0.4, 0.6],
-        [-0.55, -0.2],
-        [-1.5, -0.5],
-        [0.1, 0.2],
-        [-0.1, -0.06],
-        [0.110, 0.465],
-        [25, 125],
-        [0, 2],
-        [0.08, 0.1],
-    ]
-)
-"""array: same as prior_flat, but just as an array ordered according to standard ordering"""
-
-ordering_standard = [
-    "rlamdo",
-    "akapa",
-    "cpi",
-    "W",
-    "beto",
-    "lambda",
-    "mixed",
-    "qo3",
-    "qdirso2",
-    "qindso2",
-    "qbc",
-    "qoc",
-    "beta_f",
-    "mixed_carbon",
-    "qbmb",
-    "qh2o_ch4",
-    "threstemp",
-    "lm",
-    "ldtime",
-]
-"""list: Containing a default ordering of parameters """
 
 ordering_standard_forc = [
     "rlamdo",
@@ -85,12 +52,12 @@ ordering_standard_forc = [
     "beto",
     "lambda",
     "mixed",
+    "ocean_efficacy",
     "threstemp",
     "lm",
     "ldtime",
 ]
 """list: Containing a default ordering of parameters for forcing run"""
-aerosols = ["qdirso2", "qindso2", "qbc", "qoc"]
 
 
 class _ConfigDistro:
@@ -101,7 +68,7 @@ class _ConfigDistro:
 
     def __init__(
         self,
-        distro_array=prior_flat_array,
+        distro_dict=prior_flat,
         setvalues={
             "threstemp": 7.0,
             "lm": 40,
@@ -109,11 +76,9 @@ class _ConfigDistro:
             "qbmb": 0,
             "qh2o_ch4": 0.091915,
         },
-        ordering="standard",
         options={
             "forc": False,
             "method": "latin",
-            "aerosol-total": [-0.461590, -0.519163, 0.202893, -0.104478],
         },
     ):  # pylint:disable=dangerous-default-value
         """
@@ -121,11 +86,11 @@ class _ConfigDistro:
 
         Parameters
         ----------
-        distro_array: array
-              2D array, with first index over the parameters according
-              to the ordering given in the ordering parameter. Second index
-              of length 2, and gives the ends of the interval for the
-              parameter. If the method is latin, these are assumed to be
+        distro_array: dict
+              Listing parameters to make a config distro over. Keys should be
+              parameters to make the distribution over and keys should list
+              two numerical values that describes the configuration over which
+              to make the configurations. If the method is latin, these are assumed to be
               extent of the parameter dimension. If the method is gaussian,
               these are assumed to be points one standard deviation away
               from the center of of the distribution for the parameter
@@ -134,9 +99,6 @@ class _ConfigDistro:
         setvalues: dict
               dictionary of parameters that will have a set value, keys
               are parameter names, values are their set values
-        ordering: list
-              list of parameter ordering. This  should match the ordering
-              of values in the distro_array
         options: dict
               for optional key word arguments, such as forc, a bool which
               defines whether parameter distribution is meant for forcing
@@ -146,47 +108,37 @@ class _ConfigDistro:
               drawn from gaussian distributions for each of the
               parameters. This method will also be cosen if some other random
               string or object is sent for this keyword argument.
-              And aerosol_total, which should be a four element array
-              defining the proportions between the aerosol forcings for
-              dirso2, indso2, bc and oc in that order
         """
         self.options = options
         if "forc" not in options:
             self.options["forc"] = False
         if "method" not in options:
             self.options["method"] = "latin"
-        if "aerosol_total" in ordering and "aerosol_total" not in options:
-            options["aerosol_total"] = [-0.461590, -0.519163, 0.202893, -0.104478]
-        if "aerosol_total" in ordering:
-            self.options["aerosol_total"] = np.array(options["aerosol_total"]) / np.sum(
-                options["aerosol_total"]
-            )
-        if ordering == "standard":
-            if self.options["forc"]:
-                ordering = [
-                    o for o in ordering_standard_forc if o not in set(setvalues)
-                ]
-            else:
-                ordering = [o for o in ordering_standard if o not in set(setvalues)]
-                if "aerosol_total" in ordering:
-                    for aerosol in aerosols:
-                        ordering.remove(aerosol)
+        if self.options["forc"]:
+            ordering = [
+                o
+                for o in sorted(distro_dict.keys())
+                if (o in ordering_standard_forc) and (o not in set(setvalues))
+            ]
         else:
-            ordering = [o for o in ordering if o not in set(setvalues)]
+            ordering = [
+                o for o in sorted(distro_dict.keys()) if o not in set(setvalues)
+            ]
+
         self.ordering = ordering
-        self.prior = self._set_prior(distro_array)
+        self.prior = self._set_prior(distro_dict)
         self.setvalues = setvalues
         self._set_pamsets_start()
 
-    def _set_prior(self, distro_array):
+    def _set_prior(self, distro_dict):
         """
         Fill out the prior from a given distro_array
 
         Parameters
         ----------
-        distro_array: array
-              2D array, with first index over the parameters that according
-              to the ordering given in the ordering parameter. Second index
+        distro_dict: dict
+              Dictionary with parameters over which to make config distro.
+              Keys are parameter names and values should be numerical arrays
               of length 2, and gives the ends of the interval for the
               parameter. If the method is latin, these are assumed to be
               extent of the parameter dimension. If the method is gaussian
@@ -202,14 +154,8 @@ class _ConfigDistro:
              have setvalues according to the ordering.
         """
         prior = np.zeros([len(self.ordering), 2])
-        try:
-            len_given = len(distro_array[:, 0])
-            prior[:len_given, :] = distro_array
-        except (ValueError, TypeError):
-            LOGGER.warning("distro_array not 2 dimensional, disregarding")
-            len_given = 0
-        for i, pam in enumerate(self.ordering[len_given:]):
-            prior[len_given + i, :] = prior_flat[pam]
+        for i, pam in enumerate(self.ordering):
+            prior[i, :] = distro_dict[pam]
         return prior
 
     def _set_pamsets_start(self):
@@ -218,9 +164,12 @@ class _ConfigDistro:
         """
         self.pamset_udm_start = {}
         self.pamset_emiconc_start = {}
+        self.pamset_carbon_start = {}
         for pam, value in self.setvalues.items():
             if pam in ordering_standard_forc:
                 self.pamset_udm_start[pam] = value
+            elif pam in CARBON_CYCLE_MODEL_REQUIRED_PAMSET:
+                self.pamset_carbon_start[pam] = value
             else:
                 self.pamset_emiconc_start[pam] = value
 
@@ -301,23 +250,24 @@ class _ConfigDistro:
             samples = self.get_samples_from_distro_gaussian(numvalues)
 
         for i in range(numvalues):
-            pamset_udm = self.pamset_udm_start
-            pamset_emiconc = self.pamset_emiconc_start
+            pamset_udm = self.pamset_udm_start.copy()
+            pamset_emiconc = self.pamset_emiconc_start.copy()
+            pamset_carbon = self.pamset_carbon_start.copy()
             for j, pam in enumerate(self.ordering):
                 if pam in ordering_standard_forc:
                     pamset_udm[pam] = samples[i, j]
+                elif (
+                    not self.options["forc"]
+                    and pam in CARBON_CYCLE_MODEL_REQUIRED_PAMSET
+                ):
+                    pamset_carbon[pam] = samples[i, j]
                 elif not self.options["forc"]:
-                    if pam == "aerosol_total":
-                        for anum, aerosol in enumerate(aerosols):
-                            pamset_emiconc[aerosol] = (
-                                samples[i, j] * self.options["aerosol_total"][anum]
-                            )
-                    else:
-                        pamset_emiconc[pam] = samples[i, j]
+                    pamset_emiconc[pam] = samples[i, j]
 
             config_list[i] = {
                 "pamset_udm": pamset_udm.copy(),
                 "pamset_emiconc": pamset_emiconc.copy(),
+                "pamset_carbon": pamset_carbon.copy(),
                 "Index": f"{indexer_pre}{i}",
             }
         if json_fname != "No":