+++ tests/integration/test_calibrator.py
@@ -45,13 +45,11 @@ def test_calibrator(test_data_dir):
             "Yearstart_change": [2018, 2000],
             "Yearend_change": [2018, 2019],
             "Central Value": [320.69251537323, 0.5372],
-            "sigma": [17.020342912051203, 0.039028311931729676],
+            "sigma": [25, 0.1],
         }
     )
     # calibdata = pd.DataFrame(data = {'Variable Name': ["Heat Content|Ocean"], 'Yearstart_norm': [1971], "Yearend_norm":[1971], "Yearstart_change":[2018], "Yearend_change":[2018], "Central Value": [320.69251537323], "sigma": [17.020342912051203]})
     testconfig = _ConfigDistro(
-        distro_array=[],
-        ordering=["aerosol_total", "W", "lambda", "beta_f"],
         setvalues={
             "threstemp": 7.0,
             "lm": 40,
@@ -65,17 +63,35 @@ def test_calibrator(test_data_dir):
             "beto": 3.5,
             "mixed": 60,
         },
-        options={"aerosol_total": [-0.36, -0.97, 0.16, -0.08]},
     )
-    assert testconfig.ordering == ["aerosol_total", "W", "lambda", "beta_f"]
+    print(testconfig.ordering)
+    assert testconfig.ordering == [
+        "W",
+        "beta_f",
+        "lambda",
+        "mixed_carbon",
+        "ml_fracmax",
+        "ml_t_half",
+        "ml_w_sigmoid",
+        "npp0",
+        "ocean_efficacy",
+        "qbc",
+        "qdirso2",
+        "qindso2",
+        "qoc",
+        "solubility_limit",
+        "solubility_sens",
+        "t_half",
+        "t_threshold",
+        "w_sigmoid",
+        "w_threshold",
+    ]
     calibrator = Calibrator(calibdata, testconfig, scendata)
     drawn_cfgs = calibrator.get_n_samples(
         1, current_samples=[], kept_configs=[], recurse_num=0
     )
     assert len(drawn_cfgs) >= 1
     testconfig = _ConfigDistro(
-        distro_array=[],
-        ordering=["W", "lambda"],
         setvalues={
             "threstemp": 7.0,
             "lm": 40,
@@ -89,9 +105,10 @@ def test_calibrator(test_data_dir):
             "beto": 3.5,
             "mixed": 60,
         },
+        options={"forc": True},
     )
-    assert testconfig.ordering == ["W", "lambda"]
-    calibrator = Calibrator(calibdata, testconfig, scendata)
+    assert testconfig.ordering == ["W", "lambda", "ocean_efficacy"]
+    assert testconfig.options["method"] == "latin"
     drawn_cfgs = calibrator.get_n_samples(
         3, current_samples=[], kept_configs=[], recurse_num=0
     )