+++ tests/integration/test_parallel_run.py
@@ -114,23 +114,12 @@ def test_ciceroscm_run_parallel_many_scenarios(test_data_dir):
     )
     cfgs = [
         {
-            "pamset_udm": {
-                "rlamdo": 15.1,
-                "akapa": 0.657,
-                "cpi": 0.208,
-                "W": 2.2,
-                "beto": 6.9,
-                "lambda": 0.606,
-                "mixed": 107.0,
-            },
-            "pamset_emiconc": {
-                "qbmb": 0.0,
-                "qo3": 0.5,
-                "qdirso2": -0.3701,
-                "qindso2": -0.4163,
-                "qbc": 0.163,
-                "qoc": -0.084,
-                "qh2o_ch4": 0.171,
+            "pamset_carbon": {
+                "npp0": 60,
+                "solubility_sens": 0,
+                "ml_t_half": 10,
+                "t_half": 10,
+                "t_threshold": 10,
             },
             "Index": "13555_old_NR_rounded",
         }
@@ -217,6 +206,13 @@ def test_ciceroscm_run_parallel_many_cfgs(test_data_dir):
                 "qoc": -0.084,
                 "qh2o_ch4": 0.171,
             },
+            "pamset_carbon": {
+                "npp0": 60,
+                "solubility_sens": 0,
+                "ml_t_half": 10,
+                "t_half": 10,
+                "t_threshold": 10,
+            },
             "Index": "13555_old_NR_rounded",
         },
         {
@@ -236,6 +232,13 @@ def test_ciceroscm_run_parallel_many_cfgs(test_data_dir):
                 "qindso2": -0.96609,
                 "qbc": 0.1566,
                 "qoc": -0.0806,
+            },
+            "pamset_carbon": {
+                "npp0": 60,
+                "solubility_sens": 0,
+                "ml_t_half": 10,
+                "t_half": 10,
+                "t_threshold": 10,
             },
             "Index": "13555_old_NR_improved",
         },
@@ -257,6 +260,13 @@ def test_ciceroscm_run_parallel_many_cfgs(test_data_dir):
                 "qbc": 0.162692563111132,
                 "qoc": -0.08377713183167902,
             },
+            "pamset_carbon": {
+                "npp0": 60,
+                "solubility_sens": 0,
+                "ml_t_half": 10,
+                "t_half": 10,
+                "t_threshold": 10,
+            },
             "Index": "13555_old_NR",
         },
         {
@@ -276,6 +286,13 @@ def test_ciceroscm_run_parallel_many_cfgs(test_data_dir):
                 "qindso2": -0.3622885009449893,
                 "qbc": 0.1415852070009452,
                 "qoc": -0.07290808089507649,
+            },
+            "pamset_carbon": {
+                "npp0": 60,
+                "solubility_sens": 0,
+                "ml_t_half": 10,
+                "t_half": 10,
+                "t_threshold": 10,
             },
             "Index": "10496_old_NR",
         },
@@ -297,6 +314,13 @@ def test_ciceroscm_run_parallel_many_cfgs(test_data_dir):
                 "qbc": 0.1415852070009452,
                 "qoc": -0.07290808089507649,
             },
+            "pamset_carbon": {
+                "npp0": 60,
+                "solubility_sens": 0,
+                "ml_t_half": 10,
+                "t_half": 10,
+                "t_threshold": 10,
+            },
             "Index": "10974_old_NR",
         },
         {
@@ -317,6 +341,13 @@ def test_ciceroscm_run_parallel_many_cfgs(test_data_dir):
                 "qbc": 0.17719249872571508,
                 "qoc": -0.09124374858602939,
             },
+            "pamset_carbon": {
+                "npp0": 60,
+                "solubility_sens": 0,
+                "ml_t_half": 10,
+                "t_half": 10,
+                "t_threshold": 10,
+            },
             "Index": "Unknown_old_NR",
         },
         {
@@ -336,6 +367,13 @@ def test_ciceroscm_run_parallel_many_cfgs(test_data_dir):
                 "qindso2": -0.9929404692583796,
                 "qbc": 0.3880489761967636,
                 "qoc": -0.1998224726091362,
+            },
+            "pamset_carbon": {
+                "npp0": 60,
+                "solubility_sens": 0,
+                "ml_t_half": 10,
+                "t_half": 10,
+                "t_threshold": 10,
             },
             "Index": "28925_old_NR",
         },
@@ -418,6 +456,13 @@ def test_ciceroscm_run_parallel_many_forcing(test_data_dir):
                 "qbc": 0.163,
                 "qoc": -0.084,
                 "qh2o_ch4": 0.171,
+            },
+            "pamset_carbon": {
+                "npp0": 60,
+                "solubility_sens": 0,
+                "ml_t_half": 10,
+                "t_half": 10,
+                "t_threshold": 10,
             },
             "Index": "10496_old_NR_rounded",
         }